Clear data delete startup entry #98

+++ app/settings/src/js/index.js
@@ -135,7 +135,7 @@ const startup = () => {
 		but0.textContent = "Off"
 		startup_state = false
 
-		ipc.send("after_startup0")
+		ipc.send("disable_startup")
 	} else {
 		file.settings.launch_on_startup = true
 
@@ -144,7 +144,7 @@ const startup = () => {
 		but0.textContent = "On"
 		startup_state = true
 
-		ipc.send("after_startup1")
+		ipc.send("enable_startup")
 	}
 }
 
@@ -214,21 +214,15 @@ const reset = () => {
 							})
 
 							// remove start shortcut
-							const file_path2 = path.join(process.env.APPDATA, "/Microsoft/Windows/Start Menu/Programs/Startup/Authme Launcher.lnk")
-
 							if (dev !== true) {
-								fs.unlink(file_path2, (err) => {
-									if (err && err.code === "ENOENT") {
-										return console.warn(`Authme - Error deleting shortcut - ${err}`)
-									} else {
-										console.warn("Authme - File shortcut deleted")
-									}
-								})
+								ipc.send("disable_startup")
 							}
 
 							// clear localstorage
-							localStorage.clear()
-							sessionStorage.clear()
+							if (dev !== true) {
+								localStorage.clear()
+								sessionStorage.clear()
+							}
 
 							// restarting
 							but1.textContent = "Restarting app"
@@ -514,7 +508,7 @@ const link2 = () => {
 }
 
 const hide = () => {
-	ipc.send("hide0")
+	ipc.send("hide_settings")
 }
 
 document.querySelector(".settings").disabled = true
